Reactive Vega (2016)
According to Arvind: "Starting with VIS 2015, I think, a switch was made such that papers appearing at a given year’s VIS are published in the subsequent edition of TVCG. This was done to help with citation metrics/bean counting so that it no longer looked like VIS papers languished for their first year of publication with no citations.

So, for Reactive Vega, 2016 is the correct year and I tend to put (Proc. InfoVis ‘15) in parentheses to help clarify/bridge the confusion. "